fix: relocate SSE slow-consumer fixture for pkg-file-count
Move TestFactoryResponseEventsBySessionID_SlowConsumerFloodDoesNotBlockPublication into pkg/transports/http/servertests so pkg/transports/http stays within the 15-file package limit.
