+++ docs/internal/processes/api-relevant-files.md
@@ -343,7 +343,7 @@ dashboard state, and event connections. This agrees with
 - Keep the response-event subscription seam in `pkg/apisurface` transport-neutral: return sequence, kind, and canonical serialized JSON records rather than importing `pkg/factorysessions/responseevents` into the shared API surface. Adapt the session-owned store at the runtime edge so core packages that depend on `pkg/apisurface` do not acquire a reverse dependency on Factory Session implementation packages.
 - Retention gaps for `GET /factory-sessions/{session_id}/response-events` are synthetic `STREAM_GAP` records owned by `pkg/factorysessions/responseeventstore`: emit them before the available suffix, always pass them through response-kind filtering, report the unavailable `fromSequence`/`toSequence` bounds plus the query-relative `firstAvailableSequence`, and never publish or persist them as canonical `FactoryEvent` history.
 - Completed response-event stores accept late subscriptions for `responseeventstore.CompletedStreamRetentionWindow` (15 minutes), then return `ErrStoreExpired`; every production `apisurface.SessionAPI` implementation, including both the legacy `pkg/service` facade and initializer-backed `pkg/runtimehost.Host`, must map that domain outcome through `apisurface.ErrFactoryResponseEventStreamExpired` to the authored HTTP 410 before SSE headers are committed. Prove initializer/runtime-host composition behavior in `pkg/initializer/api_transport_internal_test.go`, not only mock or legacy service paths. When adding a typed HTTP status, also extend `errorFamilyForStatus` in `pkg/api/handlers_common.go` so the response's `family` matches its status instead of falling back to `INTERNAL_SERVER_ERROR`.
-- B08 response-event backpressure transport proofs belong in `pkg/transports/http/server_factory_response_events_slow_consumer_test.go` (`TestFactoryResponseEventsBySessionID_SlowConsumerFloodDoesNotBlockPublication` uses a gated SSE body reader so `GET /factory-sessions/{session_id}/response-events` can lag while store publication floods under a bounded budget) and `pkg/transports/cli/run/run_wire_api_test.go` (`runFactoryInvocationResponseStreamSlowStdoutFixture` covers human and NDJSON slow-stdout fixtures with `gatedResponseStreamWriter` plus terminal `invocation_result` assertions). Skip the HTTP fixture when `testing.Short()` is set.
+- B08 response-event backpressure transport proofs belong in `pkg/transports/http/servertests/server_factory_session_orchestrator_test.go` (`TestFactoryResponseEventsBySessionID_SlowConsumerFloodDoesNotBlockPublication` uses a gated SSE body reader so `GET /factory-sessions/{session_id}/response-events` can lag while store publication floods under a bounded budget) and `pkg/transports/cli/run/run_wire_api_test.go` (`runFactoryInvocationResponseStreamSlowStdoutFixture` covers human and NDJSON slow-stdout fixtures with `gatedResponseStreamWriter` plus terminal `invocation_result` assertions). Skip the HTTP fixture when `testing.Short()` is set.
 - Response-event OpenAPI contract coverage lives in `pkg/api/contracttests/openapi_contract_response_events_test.go` and `generated_contract_response_events_test.go`, with representative fixtures reused from `pkg/factorysessions/responseevents/testdata/fixtures/` plus payload/content-block coverage fixtures under `pkg/api/testdata/`. `pkg/factorysessions/responseevents` owns the canonical compact JSON envelope; because Go marshals struct fields in declaration order while generated OpenAPI structs use generator-defined ordering, keep the domain envelope and provenance order aligned with the generated server model and compare complete marshaled bytes for every shared representative fixture.
 - Internal provider/session response-stream changes stay out of the public contract by proving absence at the artifact boundary in `pkg/api/contracttests/`: when adding provider-specific normalization or internal `SessionResponseStream` semantics, add or update contract tests against `api/openapi.yaml`, `pkg/api/generated/server.gen.go`, `pkg/generatedclient/client.gen.go`, `ui/src/api/generated/openapi.ts`, and canonical public event fixtures so internal response-stream terms and provider-native event names do not leak into published schemas or `FactoryEvent` payload artifacts.
 - Historical `FactoryEvent` compatibility translation belongs at replay JSON ingestion in `pkg/replay/artifact.go` and `pkg/replay/event_stream_artifact.go`, before generated event unions discard retired fields. Keep reducers and projections canonical-only, normalize artifact and SSE inputs through the same boundary helper, and limit translation to event payload objects so customer work content is not rewritten.
